reduced compiler warning: remove -ansi flag
This causes struct timezone to be undefined, causing warning.
The reason behind this is:

Functions that are normally built in but do not have semantics defined by ISO
C (such as alloca and ffs) are not built-in functions when -ansi is used. See
Other built-in functions provided by GCC, for details of the functions affected.

https://gcc.gnu.org/onlinedocs/gcc/C-Dialect-Options.html

fixed domains:

briefcaseworld
assembly
freecell
fridge
grid
grippers
logistics
mprime
mystery
tsp
miconic
turnandopen
tyreworld