Add missing precondition in blocksworld domain (3ops)
`move-b-to-b` is missing a precondition to validate that the `?bm` and
`?bt` block are not the same block. This missing precondition allows
for a "loophole", where the following action (while not logical) is
technically valid:

(move-b-to-b b1 b2 b1)

This type of move was observed when generating solutions using
FastDownward.